% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addCountingWrapper.R
\name{addCountingWrapper}
\alias{addCountingWrapper}
\title{Return a function which counts how often it has been evaluated.}
\usage{
addCountingWrapper(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Smoof function.}
}
\value{
[\code{smoof_function}]
}
\description{
Return a function which counts how often it has been evaluated.
}

