% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doesCountEvaluations.R
\name{doesCountEvaluations}
\alias{doesCountEvaluations}
\title{Check whether the function is counting its function evaluations.}
\usage{
doesCountEvaluations(object)
}
\arguments{
\item{object}{[any]\cr
Arbitrary R object.}
}
\value{
\code{logical(1)}
 \code{TRUE} if the function counts its evaluations, \code{FALSE} otherwise.
}
\description{
In this case, the function is of type \code{smoof_counting_function} or it
is further wrapped by another wrapper. This function then checks recursively,
if there is a counting wrapper.
}
\seealso{
\code{\link{addCountingWrapper}}
}
