% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDescription.R
\name{getDescription}
\alias{getDescription}
\title{Returns the description of the function.}
\usage{
getDescription(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{character(1)}]
 A character string representing the description of the function.
}
\description{
Returns the description of the function.
}
