% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getID.R
\name{getID}
\alias{getID}
\title{Return the ID / short name of the function or \code{NA} if no ID is set.}
\usage{
getID(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{character(1)}] or \code{NA}
}
\description{
Return the ID / short name of the function or \code{NA} if no ID is set.
}
