% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocalOptimum.R
\name{getLocalOptimum}
\alias{getLocalOptimum}
\title{Returns the local optima.}
\usage{
getLocalOptimum(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{list}] List containing the following entries:
\itemize{
  \item{param [\code{list}]}{List of parameter values per local optima.}
  \item{value [\code{list}]}{List of objective values per local optima.}
  \item{is.minimum [\code{logical(1)}]}{Are the local optima minima or maxima?}
}
}
\description{
This function returns the parameters and objective values of
all local optima (including the global one).
}
\note{
Keep in mind, that this method makes sense only for single-objective target functions.
}

