% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getLoggedValues.R
\name{getLoggedValues}
\alias{getLoggedValues}
\title{Extract logged values of a function wrapped by a logging wrapper.}
\usage{
getLoggedValues(fn)
}
\arguments{
\item{fn}{[\code{smoog_function}]\cr
Smoof function.}
}
\value{
[\code{list}]
  List containing the following values
  \describe{
    \item{pars}{Data frame of parameter values, i.e., x-values or the empty
    data frame if x-values were not logged.}
    \item{obj.vals}{Numeric vector of objective vals in the single-objective
    case respectively a matrix of objective vals for multi-objective functions.}
  }
}
\description{
Extract logged values of a function wrapped by a logging wrapper.
}

