% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoxConstraints.R
\name{getLowerBoxConstraints}
\alias{getLowerBoxConstraints}
\title{Returns lower box constraints for a Smoof function.}
\usage{
getLowerBoxConstraints(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{numeric}]
 Numeric vector representing the lower box constraints
}
\description{
Returns lower box constraints for a Smoof function.
}
