% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanFunction.R
\name{getMeanFunction}
\alias{getMeanFunction}
\title{Returns the true mean function in the noisy case.}
\usage{
getMeanFunction(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{function}]
 True mean function in the noisy case.
}
\description{
Returns the true mean function in the noisy case.
}
