% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getName.R
\name{getName}
\alias{getName}
\title{Returns the name of the function.}
\usage{
getName(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{character(1)}]
 The name of the function.
}
\description{
Returns the name of the function.
}
