% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addCountingWrapper.R
\name{getNumberOfEvaluations}
\alias{getNumberOfEvaluations}
\title{Return number of function evaluations performed by the given function.}
\usage{
getNumberOfEvaluations(fn)
}
\arguments{
\item{fn}{[\code{smoof_counting_function}]\cr
Wrapped \code{smoof_function}.}
}
\value{
[\code{integer(1)}]
}
\description{
Return number of function evaluations performed by the given function.
}

