% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumberOfObjectives.R
\name{getNumberOfObjectives}
\alias{getNumberOfObjectives}
\title{Determines the number of objectives.}
\usage{
getNumberOfObjectives(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{integer(1)}]
 The number of objectives.
}
\description{
Determines the number of objectives.
}
