% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumberOfParameters.R
\name{getNumberOfParameters}
\alias{getNumberOfParameters}
\title{Determines the number of parameters.}
\usage{
getNumberOfParameters(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{integer(1)}]
 The number of parameters.
}
\description{
Determines the number of parameters.
}
