% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{getOptimaDf}
\alias{getOptimaDf}
\title{Get Data Frame of optima.}
\usage{
getOptimaDf(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Smoof function.}
}
\value{
[\code{data.frame}]
A data frame containing information about the optima.
}
\description{
This function returns data frame of global / local optima for visualization using ggplot.
}
