% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getParamSet.R
\name{getParamSet}
\alias{getParamSet}
\title{Get parameter set.}
\usage{
getParamSet(fn)
}
\arguments{
\item{fn}{[\code{otf_function}]\cr
Target function.}
}
\value{
[\code{\link[ParamHelpers]{ParamSet}}]
}
\description{
Get parameter set.
}

