% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefPoint.R
\name{getRefPoint}
\alias{getRefPoint}
\title{Returns the reference point of a multi-objective function.}
\usage{
getRefPoint(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{numeric}]
}
\description{
Returns the reference point of a multi-objective function.
}
\note{
Keep in mind, that this method makes sense only for multi-objective target functions.
}
