% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShift.R
\name{getShift}
\alias{getShift}
\title{Return the outer most shift.}
\usage{
getShift(fn, all = FALSE)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}

\item{all}{[\code{logical(1)}]\cr
Return the sum of all shifts if \code{TRUE}, else return only the outer most shift.
Default \code{FALSE}.}
}
\value{
[\code{numeric}]
}
\description{
Return the outer most shift.
}
