% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTags.R
\name{getTags}
\alias{getTags}
\title{Returns the vector of associated tags.}
\usage{
getTags(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{character}]
 Vector of associated tags.
}
\description{
Returns the vector of associated tags.
}
