% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBoxConstraints.R
\name{getUpperBoxConstraints}
\alias{getUpperBoxConstraints}
\title{Return upper box constaints.}
\usage{
getUpperBoxConstraints(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{numeric}]
}
\description{
Return upper box constaints.
}

