% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getWrappedFunction.R
\name{getWrappedFunction}
\alias{getWrappedFunction}
\title{Extract wrapped function.}
\usage{
getWrappedFunction(fn)
}
\arguments{
\item{fn}{[\code{smoof_wrapped_function}]\cr
Wrapping function.}
}
\value{
[\code{function}]
}
\description{
The \pkg{smoof} package offers means to let a function log its evaluations or
even store to points and function values it has been evaluated on. This is done
by wrapping the function with other functions. This helper function extract
the wrapped function.
}
\seealso{
\code{\link{addCountingWrapper}}, \code{\link{addLoggingWrapper}}
}

