% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hasConstraints.R
\name{hasBoxConstraints}
\alias{hasBoxConstraints}
\title{Checks whether the objective function has box constraints.}
\usage{
hasBoxConstraints(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether the objective function has box constraints.
}

