% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasGlobalOptimum.R
\name{hasGlobalOptimum}
\alias{hasGlobalOptimum}
\title{Checks whether the global optimum is known.}
\usage{
hasGlobalOptimum(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{logical(1)}]
 \code{TRUE} if the global optimum is known, \code{FALSE} otherwise.
}
\description{
Checks whether the global optimum is known.
}
