% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasGlobalOptimum.R
\name{hasLocalOptimum}
\alias{hasLocalOptimum}
\title{Checks whether local optima are known.}
\usage{
hasLocalOptimum(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether local optima are known.
}
