% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hasConstraints.R
\name{hasOtherConstraints}
\alias{hasOtherConstraints}
\title{Checks whether the objective function has other constraints.}
\usage{
hasOtherConstraints(fn)
}
\arguments{
\item{fn}{[\code{otf_function}]\cr
Target function.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether the objective function has other constraints.
}

