% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isMultiobjective}
\alias{isMultiobjective}
\title{Checks whether the given function is multiobjective.}
\usage{
isMultiobjective(fn)
}
\arguments{
\item{fn}{[\code{otf_function}]\cr
Target function.}
}
\value{
[\code{logical(1)}] \code{TRUE} if function is multiobjective.
}
\description{
Checks whether the given function is multiobjective.
}

