% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isNoisy}
\alias{isNoisy}
\title{Checks whether the given function is noisy.}
\usage{
isNoisy(fn)
}
\arguments{
\item{fn}{[\code{otf_function}]\cr
Target function.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether the given function is noisy.
}

