% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isSingleobjective}
\alias{isSingleobjective}
\title{Checks whether the given function is single-objective.}
\usage{
isSingleobjective(fn)
}
\arguments{
\item{fn}{[\code{otf_function}]\cr
Target function.}
}
\value{
[\code{logical(1)}] \code{TRUE} if function is single-objective.
}
\description{
Checks whether the given function is single-objective.
}

