% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isSmoofFunction.R
\name{isSmoofFunction}
\alias{isSmoofFunction}
\title{Checks whether the given object is a \code{smoof_function}.}
\usage{
isSmoofFunction(object)
}
\arguments{
\item{object}{[any]\cr
Arbitrary R object.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether the given object is a \code{smoof_function}.
}

