% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isVectorized.R
\name{isVectorized}
\alias{isVectorized}
\title{Checks whether the given function accepts \dQuote{vectorized} input.}
\usage{
isVectorized(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{logical(1)}]
 \code{TRUE} if the function accepts vectorized input, \code{FALSE} otherwise.
}
\description{
Checks whether the given function accepts \dQuote{vectorized} input.
}
