% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isWrappedSmoofFunction.R
\name{isWrappedSmoofFunction}
\alias{isWrappedSmoofFunction}
\title{Checks whether the function is of type \code{smoof_wrapped_function}.}
\usage{
isWrappedSmoofFunction(object)
}
\arguments{
\item{object}{[any]\cr
Arbitrary R object.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether the function is of type \code{smoof_wrapped_function}.
}

