% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.ackley.R
\name{makeAckleyFunction}
\alias{makeAckleyFunction}
\title{Ackley function.}
\usage{
makeAckleyFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{dimension}]\cr
Number of parameters.}
}
\description{
Ackley function.
}

