% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.adjiman.R
\name{makeAdjimanFunction}
\alias{makeAdjimanFunction}
\title{Adjiman function}
\usage{
makeAdjimanFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
This two-dimensional multimodal test function follows the formula
\deqn{f(\mathbf{x}) = \cos(\mathbf{x}_1)\sin(\mathbf{x}_2) - \frac{\mathbf{x}_1}{(\mathbf{x}_2^2 + 1)}}
with \eqn{\mathbf{x}_1 \in [-1, 2], \mathbf{x}_2 \in [2, 1]}.
}

