% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.alpine02.R
\name{makeAlpine02Function}
\alias{makeAlpine02Function}
\title{Alpine02 function}
\usage{
makeAlpine02Function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Alpine02 function
}

