% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.aluffi-pentini.R
\name{makeAluffiPentiniFunction}
\alias{makeAluffiPentiniFunction}
\title{Aluffi-Pentini function.}
\usage{
makeAluffiPentiniFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional test function based on the formula
\deqn{f(\mathbf{x}) = 0.25 x_1^4 - 0.5 x_1^2 + 0.1 x_1 + 0.5 x_2^2}
with \eqn{\mathbf{x}_1, \mathbf{x}_2 \in [-10, 10]}.
}
\note{
This functions is also know as the Zirilli function.
}
\references{
See \url{https://al-roomi.org/benchmarks/unconstrained/2-dimensions/26-aluffi-pentini-s-or-zirilli-s-function}.
}
