% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.amgm.R
\name{makeAMGMFunction}
\alias{makeAMGMFunction}
\title{AMGM function}
\usage{
makeAMGMFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Aritmetic Mean - Geometric Mean Equality problem. Multimodal problem defined
by the formula
\deqn{f(\mathbf{x}) = \left(\frac{1}{n} \sum_{i = 1}^{n} \mathbf{x}_i - \sqrt[n]{\prod_{i = 1}^n \mathbf{x}_i}\right)^2}
within the constraints \eqn{\mathbf{x}_i \in [0, 10]} for \eqn{i = 1, \ldots, n}.
}

