% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.bent.cigar.R
\name{makeBentCigarFunction}
\alias{makeBentCigarFunction}
\title{Bent-Cigar Function}
\usage{
makeBentCigarFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
An object of class \code{SingleObjectiveFunction}, representing the Bent-Cigar Function.

[\code{smoof_single_objective_function}]
}
\description{
Scalable test function \eqn{f} with
\deqn{f(\mathbf{x}) = x_1^2 + 10^6 \sum_{i = 2}^{n} x_i^2}
subject to \eqn{-100 \leq \mathbf{x}_i \leq 100} for \eqn{i = 1, \ldots, n}.
}
\references{
See \url{https://al-roomi.org/benchmarks/unconstrained/n-dimensions/164-bent-cigar-function}.
}
