% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.bird.R
\name{makeBirdFunction}
\alias{makeBirdFunction}
\title{Bird Function}
\usage{
makeBirdFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Multimodal single-objective test function. The implementation is based on the
mathematical formulation
\deqn{f(\mathbf{x}) = (\mathbf{x}_1 - \mathbf{x}_2)^2 + \exp((1 - \sin(\mathbf{x}_1))^2)\cos(\mathbf{x}_2) + \exp((1 - \cos(\mathbf{x}_2))^2)\sin(\mathbf{x}_1).}
The function is restricted to two dimensions with \eqn{\mathbf{x}_i \in [-2\pi, 2\pi], i = 1, 2.}
}
\references{
S. K. Mishra, Global Optimization By Differential Evolution and
Particle Swarm Methods: Evaluation On Some Benchmark Functions, Munich Research
Papers in Economics.
}
