% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.sphere.R
\name{makeBiSphereFunction}
\alias{makeBiSphereFunction}
\title{Bi-objective Sphere function}
\usage{
makeBiSphereFunction(dimensions, a = rep(0, dimensions))
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Number of decision variables.}

\item{a}{[\code{numeric(1)}]\cr
Shift parameter for the second objective. Default is (0,...,0).}
}
\value{
[\code{smoof_multi_objective_function}]
 Returns an instance of the sphere function as a \code{smoof_multi_objective_function} object.
}
\description{
Builds and returns the bi-objective Sphere test problem:
\deqn{f(\mathbf{x}) = \left(\sum_{i=1}^{n} \mathbf{x}_i^2, \sum_{i=1}^{n} (\mathbf{x}_i - \mathbf{a})^2\right)}
where \deqn{\mathbf{a} \in R^n}.
}
