% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.BK1.R
\name{makeBK1Function}
\alias{makeBK1Function}
\title{BK1 function generator}
\usage{
makeBK1Function()
}
\value{
[\code{smoof_multi_objective_function}]
 Returns an instance of the BK1 function as a \code{smoof_multi_objective_function} object.
}
\description{
Generates the BK1 function, a multi-objective optimization test function. The BK1 function
is commonly used in benchmarking studies for evaluating the performance of optimization algorithms.
}
\references{
...
}
