% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.booth.R
\name{makeBoothFunction}
\alias{makeBoothFunction}
\title{Booth Function}
\usage{
makeBoothFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
This function is based on the formula
\deqn{f(\mathbf{x}) = (\mathbf{x}_1 + 2\mathbf{x}_2 - 7)^2 + (2\mathbf{x}_1 + \mathbf{x}_2 - 5)^2}
subject to \eqn{\mathbf{x}_i \in [-10, 10], i = 1, 2}.
}

