% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.bukin.n2.R
\name{makeBukinN2Function}
\alias{makeBukinN2Function}
\title{Bukin function N. 2}
\usage{
makeBukinN2Function()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Muldimodal, non-scalable, continous optimization test function given by:
\deqn{f(\mathbf{x}) = 100 (\mathbf{x}_2 - 0.01 * \mathbf{x}_1^2 + 1) + 0.01 (\mathbf{x}_1 + 10)^2}
subject to \eqn{\mathbf{x}_1 \in [-15, 5]} and \eqn{\mathbf{x}_2 \in [-3, 3]}.
}
\references{
Z. K. Silagadze, Finding Two-Dimesnional Peaks, Physics of Particles
and Nuclei Letters, vol. 4, no. 1, pp. 73-80, 2007.
}
\seealso{
\code{\link{makeBukinN4Function}}, \code{\link{makeBukinN6Function}}
}

