% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.bukin.n4.R
\name{makeBukinN4Function}
\alias{makeBukinN4Function}
\title{Bukin function N. 4}
\usage{
makeBukinN4Function()
}
\value{
An object of class \code{SingleObjectiveFunction}, representing the Bukin N.4 Function.

[\code{smoof_single_objective_function}]
}
\description{
Second continuous Bukin function test function. The formula is given by
\deqn{f(\mathbf{x}) = 100 \mathbf{x}_2^2 + 0.01 * ||\mathbf{x}_1 +10||}
and the box constraints \eqn{mathbf{x}_1 \in [-15, 5], \mathbf{x}_2 \in [-3, 3]}.
}
\references{
Z. K. Silagadze, Finding Two-Dimesnional Peaks, Physics of Particles
and Nuclei Letters, vol. 4, no. 1, pp. 73-80, 2007.
}
\seealso{
\code{\link{makeBukinN2Function}}, \code{\link{makeBukinN6Function}}
}
