% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.bukin.n6.R
\name{makeBukinN6Function}
\alias{makeBukinN6Function}
\title{Bukin function N. 6}
\usage{
makeBukinN6Function()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Bukin function N. 6
}

