% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sof.carrom.table.R
\name{makeCarromTableFunction}
\alias{makeCarromTableFunction}
\title{Carrom Table Function}
\usage{
makeCarromTableFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
This function is defined as follows:
\deqn{f(\mathbf{x}) = -\frac{1}{30} \left((\cos(\mathbf{x}_1)\exp(|1 - ((\mathbf{x}_1^2 + \mathbf{x}_2^2)^{0.5} / \pi)^2)|\right).}
The box-constraints are given by \eqn{\mathbf{x}_i \in [-10, 10], i = 1, 2}.
}
\references{
S. K. Mishra, Global Optimization By Differential Evolution and
Particle Swarm Methods: Evaluation On Some Benchmark Functions, Munich
Research Papers in Economics.
}

