% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.complex.R
\name{makeComplexFunction}
\alias{makeComplexFunction}
\title{Complex function.}
\usage{
makeComplexFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional test function based on the formula
\deqn{f(\mathbf{x}) = (x_1^3 - 3 x_1 x_2^2 - 1)^2 + (3 x_2 x_1^2 - x_2^3)^2}
with \eqn{\mathbf{x}_1, \mathbf{x}_2 \in [-2, 2]}.
}
\references{
See \url{http://al-roomi.org/benchmarks/unconstrained/2-dimensions/43-complex-function}.
}

