% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.cross.in.tray.R
\name{makeCrossInTrayFunction}
\alias{makeCrossInTrayFunction}
\title{CrossInTray Function}
\usage{
makeCrossInTrayFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
CrossInTray Function
}

