% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.deflected.corrugated.spring.R
\name{makeDeflectedCorrugatedSpringFunction}
\alias{makeDeflectedCorrugatedSpringFunction}
\title{Deflected Corrugated Spring function}
\usage{
makeDeflectedCorrugatedSpringFunction(dimensions, K = 5, alpha = 5)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}

\item{K}{[\code{numeric(1)}]\cr
Parameter.}

\item{alpha}{[\code{numeric(1)}]\cr
Parameter.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Deflected Corrugated Spring function
}

