% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.dent.R
\name{makeDentFunction}
\alias{makeDentFunction}
\title{Dent Function}
\usage{
makeDentFunction()
}
\value{
[\code{smoof_multi_objective_function}]
}
\description{
Builds and returns the bi-objective Dent test problem, which is defined as
follows:
\deqn{f(\mathbf{x}) = \left(f_1(\mathbf{x}_1), f_2(\mathbf{x})\right)}
with
\deqn{f_1(\mathbf{x}_1) = 0.5 \left( \sqrt(1 + (x_1 + x_2)^2) + \sqrt(1 + (x_1 - x_2)^2) + x_1 - x_2\right) + d}
and
\deqn{f_1(\mathbf{x}_1) = 0.5 \left( \sqrt(1 + (x_1 + x_2)^2) + \sqrt(1 + (x_1 - x_2)^2) - x_1 + x_2\right) + d}
where \eqn{d = \lambda * \exp(-(x_1 - x_2)^2)} and \eqn{\mathbf{x}_i \in [-1.5, 1.5], i = 1, 2}.
}
