% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.drop.wave.R
\name{makeDropWaveFunction}
\alias{makeDropWaveFunction}
\title{Drop-Wave function}
\usage{
makeDropWaveFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Drop-Wave function
}

