% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mof.dtlz2.R
\name{makeDTLZ2Function}
\alias{makeDTLZ2Function}
\title{DTLZ2 function (family) generator.}
\usage{
makeDTLZ2Function(dimensions, n.objectives)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Number of decision variables.}

\item{n.objectives}{[\code{integer(1)}]\cr
Number of objectives.}
}
\value{
[\code{smoof_multi_objective_function}]
}
\description{
Builds and returns the multi-objective DTLZ2 test problem.
}
\references{
K. Deb and L. Thiele and M. Laumanns and E. Zitzler. Scalable
Multi-Objective Optimization Test Problems.
}

