% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mof.dtlz4.R
\name{makeDTLZ4Function}
\alias{makeDTLZ4Function}
\title{DTLZ4 function (family) generator.}
\usage{
makeDTLZ4Function(dimensions, n.objectives, alpha = 100)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Number of decision variables.}

\item{n.objectives}{[\code{integer(1)}]\cr
Number of objectives.}

\item{alpha}{[\code{numeric(1)}]\cr
Optional parameter. Default is 100, which is recommended by Deb et al.}
}
\value{
[\code{smoof_multi_objective_function}]
}
\description{
Builds and returns the multi-objective DTLZ4 test problem. It is a slight
modification of the DTLZ2 problems by introducing the parameter \eqn{\alpha}.
The parameter is used to map \eqn{\mathbf{x}_i \rightarrow \mathbf{x}_i^{\alpha}}.
}
\references{
K. Deb and L. Thiele and M. Laumanns and E. Zitzler. Scalable
Multi-Objective Optimization Test Problems.
}

