% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.eggcrate.R
\name{makeEggCrateFunction}
\alias{makeEggCrateFunction}
\title{Egg crate function}
\usage{
makeEggCrateFunction()
}
\value{
[\code{otf_single_objective_function}]
}
\description{
Egg crate function
}

