% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.eggholder.R
\name{makeEggholderFunction}
\alias{makeEggholderFunction}
\title{Egg Holder function}
\usage{
makeEggholderFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Egg Holder function
}

