% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.el.attar.vidyasagar.dutta.R
\name{makeElAttarVidyasagarDuttaFunction}
\alias{makeElAttarVidyasagarDuttaFunction}
\title{ElAttarVidyasagarDutta function}
\usage{
makeElAttarVidyasagarDuttaFunction()
}
\description{
ElAttarVidyasagarDutta function
}

