% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.engvall.R
\name{makeEngvallFunction}
\alias{makeEngvallFunction}
\title{Complex function.}
\usage{
makeEngvallFunction()
}
\value{
An object of class \code{SingleObjectiveFunction}, representing the Complex Function.

[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional test function based on the formula
\deqn{f(\mathbf{x}) = (x_1^4 + x_2^4 + 2 x_1^2 x_2^2 - 4 x_1 + 3}
with \eqn{\mathbf{x}_1, \mathbf{x}_2 \in [-2000, 2000]}.
}
\references{
See \url{https://al-roomi.org/benchmarks/unconstrained/2-dimensions/116-engvall-s-function}.
}
