% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.freudenstein.roth.R
\name{makeFreudensteinRothFunction}
\alias{makeFreudensteinRothFunction}
\title{Freudenstein Roth Function}
\usage{
makeFreudensteinRothFunction()
}
\value{
[\code{otf_single_objective_function}]
}
\description{
Freudenstein Roth Function
}

