% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeFunctionByName.R
\name{makeFunctionByName}
\alias{makeFunctionByName}
\title{Convenience function to generate a smoof function by passing the function name
as a string.}
\usage{
makeFunctionByName(fun.name, ...)
}
\arguments{
\item{fun.name}{[\code{character(1)}]\cr
Name of the generator function as a string.}

\item{...}{[any]\cr
Further arguments passed to generator.}
}
\value{
[\code{smoof_function}]
}
\description{
Convenience function to generate a smoof function by passing the function name
as a string.
}

