% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.hartmann3.R
\name{makeHartmann3Function}
\alias{makeHartmann3Function}
\title{Hartmann 3 Function}
\usage{
makeHartmann3Function()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Unimodal single-objective test function with three local minima. 
The implementation is based on the mathematical formulation
\deqn{f(x) = - \sum_{i=1}^4 \alpha_i \ exp \left(-\sum_{j=1}^3 A_{ij}(x_j-P_{ij})^2 \right)}, where 
\deqn{\alpha = (1.0, 1.2, 3.0, 3.2)^T, \\
 A = \left( \begin{array}{rrrrrr}
  3.0 & 10 & 30 \\ 
  0.1 & 10 & 35 \\ 
  3.0 & 10 & 30 \\ 
  0.1 & 10 & 35 \\ 
 \end{array} \right), \\
  P = 10^{-4} \cdot \left(\begin{array}{rrr}
  3689 & 1170 & 2673 \\ 
  4699 & 4387 & 7470 \\ 
  1091 & 8732 & 5547 \\ 
   381 & 5743 & 8828
 \end{array} \right)}
The function is restricted to three dimensions with \eqn{\mathbf{x}_i \in [0,1], i = 1, \ldots, 3.}
}
\references{
Picheny, V., Wagner, T., & Ginsbourger, D. (2012). A benchmark 
of kriging-based infill criteria for noisy optimization.
}

