% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.hartmann4.R
\name{makeHartmann4Function}
\alias{makeHartmann4Function}
\title{Hartmann 4 Function}
\usage{
makeHartmann4Function()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Unimodal single-objective test function with four local minima. 
It is designed to have a mean of zero and a variance of one.
The implementation is based on the mathematical formulation
\deqn{f(x) = - \sum_{i=1}^4 \alpha_i \ exp \left(-\sum_{j=1}^4 A_{ij}(x_j-P_{ij})^2 \right)}, where 
\deqn{\alpha = (1.0, 1.2, 3.0, 3.2)^T, \\
 A = \left( \begin{array}{rrrrrr}
  10   & 3    & 17   & 3.50 \\
  0.05 & 10   & 17   & 0.1  \\
  3    & 3.5  & 1.7  & 10   \\
  17   & 8    & 0.05 & 10   
 \end{array} \right), \\
  P = 10^{-4} \cdot \left(\begin{array}{rrrrrr}
  1312 & 1696 & 5569 & 124  \\
  2329 & 4135 & 8307 & 3736 \\
  2348 & 1451 & 3522 & 2883 \\
  4047 & 8828 & 8732 & 5743 
 \end{array} \right)}
The function is restricted to four dimensions with \eqn{\mathbf{x}_i \in [0,1], i = 1, \ldots, 4.}
}
\references{
Picheny, V., Wagner, T., & Ginsbourger, D. (2012). A benchmark 
of kriging-based infill criteria for noisy optimization.
}

