% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.himmelblau.R
\name{makeHimmelblauFunction}
\alias{makeHimmelblauFunction}
\title{Himmelblau Function}
\usage{
makeHimmelblauFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional test function based on the function defintion
\deqn{f(\mathbf{x}) = (\mathbf{x}_1^2 + \mathbf{x}_2 - 11)^2 + (\mathbf{x}_1 + \mathbf{x}_2^2 - 7)^2}
with box-constraings \eqn{\mathbf{x}_i \in [-5, 5], i = 1, 2}.
}
\references{
D. M. Himmelblau, Applied Nonlinear Programming, McGraw-Hill, 1972.
}
