% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.holder.table.R
\name{makeHolderTableFunction}
\alias{makeHolderTableFunction}
\title{Holder Table function}
\usage{
makeHolderTableFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Holder Table function
}

