% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.holder.table.n1.R
\name{makeHolderTableN1Function}
\alias{makeHolderTableN1Function}
\title{Holder Table function N. 1}
\usage{
makeHolderTableN1Function()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
This multimodal function is defined as
\deqn{f(\mathbf{x}) = -\left|\cos(\mathbf{x}_1)\cos(\mathbf{x}_2)\exp(|1 - \sqrt{\mathbf{x}_1 + \mathbf{x}_2}/\pi|)\right|}
with box-constraints \eqn{\mathbf{x}_i \in [-10, 10], i = 1, 2}.
}
\references{
S. K. Mishra, Global Optimization By Differential Evolution and
Particle Swarm Methods: Evaluation On Some Benchmark Functions, Munich
Research Papers in Economics.
}
\seealso{
\code{\link{makeHolderTableN2Function}}
}
