% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.hosaki.R
\name{makeHosakiFunction}
\alias{makeHosakiFunction}
\title{Hosaki Function}
\usage{
makeHosakiFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional test function \eqn{f} with
\deqn{f(\mathbf{x}) = (1 - 8 \mathbf{x}_1 + 7 \mathbf{x}_1^2 - 7/3 \mathbf{x}_1^3 + 1/4 \mathbf{x}_1^4)\mathbf{x}_2^2e^{-\mathbf{x}_2}}
subject to \eqn{0 \leq \mathbf{x}_1 \leq 5} and \eqn{0 \leq \mathbf{x}_2 \leq 6}.
}
\references{
G. A. Bekey, M. T. Ung, A Comparative Evaluation of Two Global
Search Algorithms, IEEE Transaction on Systems, Man and Cybernetics, vol. 4,
no. 1, pp. 112- 116, 1974.
}

