% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.vincent.R
\name{makeInvertedVincentFunction}
\alias{makeInvertedVincentFunction}
\title{Inverted Vincent Function}
\usage{
makeInvertedVincentFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
An object of class \code{SingleObjectiveFunction}, representing the Inverted Vincent Function.

[\code{smoof_single_objective_function}]
}
\description{
Single-objective test function based on the formula
\deqn{f(\mathbf{x}) = \frac{1}{n} \sum_{i = 1}^{n} \sin(10 \log(\mathbf{x}_i))}
subject to \eqn{\mathbf{x}_i \in [0.25, 10]} for \eqn{i = 1, \ldots, n}.
}
\references{
Xiadong Li, Andries Engelbrecht, and Michael G. Epitropakis. Benchmark
functions for CEC2013 special session and competition on niching methods for
multi-modal function optimization. Technical report, RMIT University, Evolutionary
Computation and Machine Learning Group, Australia, 2013.
}
