% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.jennrichsampson.R
\name{makeJennrichSampsonFunction}
\alias{makeJennrichSampsonFunction}
\title{Jennrich-Sampson function.}
\usage{
makeJennrichSampsonFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional test function based on the formula
\deqn{f(\mathbf{x}) = \sum_{i=1}^{10} \left[2 + 2i - (e^{ix_1} + e^{ix_2})\right]^2}
with \eqn{\mathbf{x}_1, \mathbf{x}_2 \in [-1, 1]}.
}
\references{
See \url{http://al-roomi.org/benchmarks/unconstrained/2-dimensions/134-jennrich-sampson-s-function}.
}

