% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.judge.R
\name{makeJudgeFunction}
\alias{makeJudgeFunction}
\title{Judge function.}
\usage{
makeJudgeFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional test function based on the formula
\deqn{f(\mathbf{x}) = \sum_{i=1}^{20} \left[(x_1 + B_i x_2 + C_i x_2^2) - A_i\right]^2}
with \eqn{\mathbf{x}_1, \mathbf{x}_2 \in [-10, 10]}. For details on \eqn{A, B, C}
see the referenced website.
}
\references{
See \url{http://al-roomi.org/benchmarks/unconstrained/2-dimensions/133-judge-s-function}.
}
