% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.katsuura.R
\name{makeKatsuuraFunction}
\alias{makeKatsuuraFunction}
\title{Katsuura function}
\usage{
makeKatsuuraFunction(dimensions, d = 32)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}

\item{d}{[\code{integer(1)}]\cr
Parameter.}
}
\value{
[\code{otf_single_objective_function}]
}
\description{
Katsuura function
}

