% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.keane.R
\name{makeKeaneFunction}
\alias{makeKeaneFunction}
\title{Keane Function}
\usage{
makeKeaneFunction()
}
\value{
An object of class \code{SingleObjectiveFunction}, representing the Keane Function.

[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional test function based on the definition
\deqn{f(\mathbf{x}) = \frac{\sin^2(\mathbf{x}_1 - \mathbf{x}_2)\sin^2(\mathbf{x}_1 + \mathbf{x}_2)}{\sqrt{\mathbf{x}_1^2 + \mathbf{x}_2^2}}.}
The domain of definition is bounded by the box constraints
\eqn{\mathbf{x}_i \in [0, 10], i = 1, 2}.
}
