% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.kearfott.R
\name{makeKearfottFunction}
\alias{makeKearfottFunction}
\title{Kearfott function.}
\usage{
makeKearfottFunction()
}
\value{
An object of class \code{SingleObjectiveFunction}, representing the Kearfott Function.

[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional test function based on the formula
\deqn{f(\mathbf{x}) = (x_1^2 + x_2^2 - 2)^2 + (x_1^2 - x_2^2 - 1)^2}
with \eqn{\mathbf{x}_1, \mathbf{x}_2 \in [-3, 4]}.
}
\references{
See \url{https://al-roomi.org/benchmarks/unconstrained/2-dimensions/59-kearfott-s-function}.
}
