% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.matyas.R
\name{makeMatyasFunction}
\alias{makeMatyasFunction}
\title{Matyas Function}
\usage{
makeMatyasFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional, unimodal test function
\deqn{f(\mathbf{x}) = 0.26 (\mathbf{x}_1^2 + \mathbf{x}_2^2) - 0.48\mathbf{x}_1\mathbf{x}_2}
subject to \eqn{\mathbf{x}_i \in [-10, 10], i = 1, 2}.
}
\references{
A.-R. Hedar, Global Optimization Test Problems.
}
