% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.max.mod.R
\name{makeMaxModFunction}
\alias{makeMaxModFunction}
\title{MaxMod function.}
\usage{
makeMaxModFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{otf_single_objective_function}]
}
\description{
MaxMod function.
}

