% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.mccormick.R
\name{makeMcCormickFunction}
\alias{makeMcCormickFunction}
\title{McCormick Function}
\usage{
makeMcCormickFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional, multimodal test function. The defintion is given by
\deqn{f(\mathbf{x}) = \sin(\mathbf{x}_1 + \mathbf{x}_2) + (\mathbf{x}_1 - \mathbf{x}_2)^2 - 1.5 \mathbf{x}_1 + 2.5 \mathbf{x}_2 + 1}
subject to \eqn{\mathbf{x}_1 \in [-1.5, 4], \mathbf{x}_2 \in [-3, 3]}.
}
\references{
F. A. Lootsma (ed.), Numerical Methods for Non-Linear
Optimization, Academic Press, 1972.
}

