% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.michalewicz.R
\name{makeMichalewiczFunction}
\alias{makeMichalewiczFunction}
\title{Michalewicz function}
\usage{
makeMichalewiczFunction(m = 5)
}
\arguments{
\item{m}{[\code{integer(1)}]\cr
Parameter.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Michalewicz function
}

