% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.mmf12.R
\name{makeMMF12Function}
\alias{makeMMF12Function}
\title{MMF12 Function}
\usage{
makeMMF12Function(np = 2L, q = 4L)
}
\arguments{
\item{np}{[\code{integer}(1)]\cr
Number of global Pareto sets. In the CEC2019 competition, the organizers used
\code{np = 2L}.}

\item{q}{[\code{integer}(1)]\cr
Number of discontinuous pieces in each Pareto front. In the CEC2019 competition,
the organizers used \code{q = 4L}.}
}
\value{
[\code{smoof_multi_objective_function}]
}
\description{
Test problem from the set of "multimodal multiobjective functions" as for
instance used in the CEC2019 competition.
}
\references{
Caitong Yue, Boyang Qu, Kunjie Yu, Jing Liang, and Xiaodong Li, "A novel
scalable test problem suite for multimodal multiobjective optimization," in
Swarm and Evolutionary Computation, Volume 48, August 2019, pp. 62–71, Elsevier.
}
