% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.mmf14.R
\name{makeMMF14Function}
\alias{makeMMF14Function}
\title{MMF14 Function}
\usage{
makeMMF14Function(dimensions, n.objectives, np = 2L)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Number of decision variables.}

\item{n.objectives}{[\code{integer(1)}]\cr
Number of objectives.}

\item{np}{[\code{integer}(1)]\cr
Number of global Pareto sets. In the CEC2019 competition, the organizers used
\code{np = 2L}.}
}
\value{
[\code{smoof_multi_objective_function}]
Returns an instance of the MMF14 function as a \code{smoof_multi_objective_function} object.
}
\description{
Test problem from the set of "multi-modal multi-objective functions" as for
instance used in the CEC2019 competition.
}
\references{
Caitong Yue, Boyang Qu, Kunjie Yu, Jing Liang, and Xiaodong Li, "A novel
scalable test problem suite for multi-modal multi-objective optimization," in
Swarm and Evolutionary Computation, Volume 48, August 2019, pp. 62–71, Elsevier.
}
