% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.mmf1e.R
\name{makeMMF1eFunction}
\alias{makeMMF1eFunction}
\title{MMF1e Function}
\usage{
makeMMF1eFunction(a = exp(1L))
}
\arguments{
\item{a}{[\code{double}(1)]\cr
Parametrizable factor. In the CEC2019 competition, the organizers used
\code{a = exp(1L)}.}
}
\value{
[\code{smoof_multi_objective_function}]
 Returns an instance of the MMF1e function as a \code{smoof_multi_objective_function} object.
}
\description{
Test problem from the set of "multi-modal multi-objective functions" as for
instance used in the CEC2019 competition.
}
\references{
Caitong Yue, Boyang Qu, Kunjie Yu, Jing Liang, and Xiaodong Li, "A novel
scalable test problem suite for multimodal multiobjective optimization," in
Swarm and Evolutionary Computation, Volume 48, August 2019, pp. 62–71, Elsevier.
}
