% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.mmf1.R
\name{makeMMF1Function}
\alias{makeMMF1Function}
\title{MMF1 Function}
\usage{
makeMMF1Function()
}
\value{
[\code{smoof_multi_objective_function}]
 Returns an instance of the MMF1 function as a \code{smoof_multi_objective_function} object.
}
\description{
Test problem from the set of "multi-modal multi-objective functions" as for
instance used in the CEC2019 competition.
}
\references{
Caitong Yue, Boyang Qu, Kunjie Yu, Jing Liang, and Xiaodong Li, "A novel
scalable test problem suite for multimodal multiobjective optimization," in
Swarm and Evolutionary Computation, Volume 48, August 2019, pp. 62–71, Elsevier.
}
