% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.mmf4.R
\name{makeMMF4Function}
\alias{makeMMF4Function}
\title{MMF4 Function}
\usage{
makeMMF4Function()
}
\value{
[\code{smoof_multi_objective_function}]
}
\description{
Test problem from the set of "multimodal multiobjective functions" as for
instance used in the CEC2019 competition.
}
\references{
Caitong Yue, Boyang Qu, Kunjie Yu, Jing Liang, and Xiaodong Li, "A novel
scalable test problem suite for multimodal multiobjective optimization," in
Swarm and Evolutionary Computation, Volume 48, August 2019, pp. 62–71, Elsevier.
}
