% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.MOP1.R
\name{makeMOP1Function}
\alias{makeMOP1Function}
\title{MOP1 function generator.}
\usage{
makeMOP1Function()
}
\value{
[\code{smoof_multi_objective_function}]
}
\description{
MOP1 function from Van Valedhuizen's test suite.
}
\references{
J. D. Schaffer, "Multiple objective optimization with vector evaluated
genetic algorithms," in Proc. 1st Int. Conf. Genetic Algorithms and Their
Applications, J. J. Grenfenstett, Ed., 1985, pp. 93-100.
}

