% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.MOP2.R
\name{makeMOP2Function}
\alias{makeMOP2Function}
\title{MOP2 function generator.}
\usage{
makeMOP2Function(dimensions = 2L)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_multi_objective_function}]
}
\description{
MOP2 function from Van Valedhuizen's test suite due to Fonseca and Fleming.
}
\references{
C. M. Fonseca and P. J. Fleming, "Multiobjective genetic algorithms
made easy: Selection, sharing and mating restriction," Genetic Algorithms
in Engineering Systems: Innovations and Applications, pp. 45-52, Sep. 1995. IEE.
}

