% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.MOP3.R
\name{makeMOP3Function}
\alias{makeMOP3Function}
\title{MOP3 function generator.}
\usage{
makeMOP3Function(dimensions = 2L)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_multi_objective_function}]
}
\description{
MOP3 function from Van Valedhuizen's test suite.
}
\references{
C. Poloni, G. Mosetti, and S. Contessi, "Multi objective optimization by
GAs: Application to system and component design," in Proc. Comput.
Methods in Applied Sciences'96: Invited Lectures and Special Technological
Sessions of the 3rd ECCOMAS Comput. Fluid Dynamics Conf.
and the 2nd ECCOMAS Conf. Numerical Methods in Engineering, Sep.
1996, pp. 258-264
}
