% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.MOP4.R
\name{makeMOP4Function}
\alias{makeMOP4Function}
\title{MOP4 function generator.}
\usage{
makeMOP4Function()
}
\value{
[\code{smoof_multi_objective_function}]
 Returns an instance of the MOP4 function as a \code{smoof_multi_objective_function} object.
}
\description{
MOP4 function from Van Valedhuizen's test suite based on Kursawe.
}
\references{
F. Kursawe, "A variant of evolution strategies for vector optimization,"
in Lecture Notes in Computer Science, H.-P. Schwefel and R. Maenner,
Eds. Berlin, Germany: Springer-Verlag, 1991, vol. 496, Proc. Parallel
Problem Solving From Nature. 1st Workshop, PPSN I, pp. 193-197.
}
