% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.MOP5.R
\name{makeMOP5Function}
\alias{makeMOP5Function}
\title{MOP5 function generator.}
\usage{
makeMOP5Function()
}
\value{
[\code{smoof_multi_objective_function}]
 Returns an instance of the MOP5 function as a \code{smoof_multi_objective_function} object.
}
\description{
MOP5 function from Van Valedhuizen's test suite.
}
\note{
Original box constraints where \eqn{[-3, 3]}.
}
\references{
R. Viennet, C. Fonteix, and I. Marc, "Multi criteria optimization using a
genetic algorithm for determining a Pareto set," Int. J. Syst. Sci., vol. 27,
no. 2, pp. 255-260, 1996
}
