% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.MOP6.R
\name{makeMOP6Function}
\alias{makeMOP6Function}
\title{MOP6 function generator.}
\usage{
makeMOP6Function()
}
\value{
[\code{smoof_multi_objective_function}]
 Returns an instance of the MOP6 function as a \code{smoof_multi_objective_function} object.
}
\description{
MOP6 function from Van Valedhuizen's test suite.
}
