% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.multi.mod.R
\name{makeMultiModFunction}
\alias{makeMultiModFunction}
\title{MultiMod function.}
\usage{
makeMultiModFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
MultiMod function.
}

