% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.periodic.R
\name{makePeriodicFunction}
\alias{makePeriodicFunction}
\title{Periodic Function}
\usage{
makePeriodicFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Single-objective two-dimensional test function. The formula is given as
\deqn{f(\mathbf{x}) = 1 + \sin^2(\mathbf{x}_1) + \sin^2(\mathbf{x}_2) - 0.1e^{-(\mathbf{x}_1^2 + \mathbf{x}_2^2)}}
subject to the constraints \eqn{\mathbf{x}_i \in [-10, 10], i = 1, 2.}
}
\references{
M. M. Ali, C. Khompatraporn, Z. B. Zabinsky, A Numerical
Evaluation of Several Stochastic Algorithms on Selected Continuous Global
Optimization Test Problems, Journal of Global Optimization, vol. 31, pp.
635-672, 2005.
}
