% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.powell.sum.R
\name{makePowellSumFunction}
\alias{makePowellSumFunction}
\title{Powell-Sum Function}
\usage{
makePowellSumFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
An object of class \code{SingleObjectiveFunction}, representing the Powell-Sum Function.

[\code{smoof_single_objective_function}]
}
\description{
The formula that underlies the implementation is given by
\deqn{f(\mathbf{x}) = \sum_{i=1}^n |\mathbf{x}_i|^{i+1}}
with \eqn{\mathbf{x}_i \in [-1, 1], i = 1, \ldots, n}.
}
\references{
S. Rahnamyan, H. R. Tizhoosh, N. M. M. Salama, A Novel Population
Initialization Method for Accelerating Evolutionary Algorithms, Computers and
Mathematics with Applications, vol. 53, no. 10, pp. 1605-1614, 2007.
}
