% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sof.price.n1.R
\name{makePriceN1Function}
\alias{makePriceN1Function}
\title{Price Function N. 1}
\usage{
makePriceN1Function()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Second function by Price. The implementation is based on the defintion
\deqn{f(\mathbf{x}) = (|\mathbf{x}_1| - 5)^2 + (|\mathbf{x}_2 - 5)^2}
subject to \eqn{\mathbf{x}_i \in [-500, 500], i = 1, 2}.
}
\references{
W. L. Price, A Controlled Random Search Procedure for Global
Optimisation, Computer journal, vol. 20, no. 4, pp. 367-370, 1977.
}
\seealso{
\code{\link{makePriceN2Function}}, \code{\link{makePriceN4Function}}
}

