% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.price.n2.R
\name{makePriceN2Function}
\alias{makePriceN2Function}
\title{Price Function N. 2}
\usage{
makePriceN2Function()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Second function by Price. The implementation is based on the defintion
\deqn{f(\mathbf{x}) = 1 + \sin^2(\mathbf{x}_1) + \sin^2(\mathbf{x}_2) - 0.1 \exp(-\mathbf{x}^2 - \mathbf{x}_2^2)}
subject to \eqn{\mathbf{x}_i \in [-10, 10], i = 1, 2}.
}
\references{
W. L. Price, A Controlled Random Search Procedure for Global
Optimisation, Computer journal, vol. 20, no. 4, pp. 367-370, 1977.
}
\seealso{
\code{\link{makePriceN1Function}}, \code{\link{makePriceN4Function}}
}
