% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.price.n3.R
\name{makePriceN3Function}
\alias{makePriceN3Function}
\title{Price Function N. 3}
\usage{
makePriceN3Function()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Third function by Price. The implementation is based on the defintion
\deqn{f(\mathbf{x}) = 100 (\mathbf{x}_2 - \mathbf{x}_1^2)^2 + \left[6.4 (\mathbf{x}_2 - 0.5)^2 - \mathbf{x}_1 - 0.6\right]^2}
subject to \eqn{\mathbf{x}_i \in [-500, 500]}.
}
\references{
W. L. Price, A Controlled Random Search Procedure for Global
Optimisation, Computer journal, vol. 20, no. 4, pp. 367-370, 1977.
}
\seealso{
\code{\link{makePriceN1Function}}, \code{\link{makePriceN2Function}},
\code{\link{makePriceN4Function}}
}

