% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.price.n4.R
\name{makePriceN4Function}
\alias{makePriceN4Function}
\title{Price Function N. 4}
\usage{
makePriceN4Function()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Fourth function by Price. The implementation is based on the defintion
\deqn{f(\mathbf{x}) = (2\mathbf{x}_1^3 \mathbf{x}_2 - \mathbf{x}_2^3)^2 + (6\mathbf{x}_1 - \mathbf{x}_2^2 + \mathbf{x}_2)^2}
subject to \eqn{\mathbf{x}_i \in [-500, 500]}.
}
\references{
W. L. Price, A Controlled Random Search Procedure for Global
Optimisation, Computer journal, vol. 20, no. 4, pp. 367-370, 1977.
}
\seealso{
\code{\link{makePriceN1Function}}, \code{\link{makePriceN2Function}}
}
