% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.rastrigin.R
\name{makeRastriginFunction}
\alias{makeRastriginFunction}
\title{Rastrigin function}
\usage{
makeRastriginFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
One of the most popular single-objective test functions consists of many
local optima and is thus highly multimodal with a global structure.
The implementation follows the formula
\deqn{f(\mathbf{x}) = 10n + \sum_{i=1}^{n} \left(\mathbf{x}_i^2 - 10 \cos(2\pi \mathbf{x}_i)\right).}
}

