% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.nk.R
\name{makeRMNKFunction}
\alias{makeRMNKFunction}
\title{rMNK-generator}
\usage{
makeRMNKFunction(M, N, K, rho = 0, funs = NULL)
}
\arguments{
\item{M}{[\code{integer(1)}]\cr
Number of objectives (at least two).}

\item{N}{[\code{integer(1)}]\cr
Length of the bit-string (decision space dimension).}

\item{K}{[\code{integer}]\cr
Integer vector of the number of epistatic interactions.
If a single value is passed a homogeneous NK-landscape is generated, i.e.
\eqn{k_i = k \forall i = 1, \ldots, N}.}

\item{rho}{[\code{numeric(1)}]\cr
Correlation between objectives.
Defaults to zero, i.e., no correlation at all.}

\item{funs}{[\code{list} | \code{NULL}]\cr
Allows for an alternative way to build a MNK-landscape by passing a list of
at least two single-objective NK-landscapes. In this case all other parameters
are ignored. Note that the passed function must be compatible, i.e., the
input dimension \code{N} needs to match.
Default is \code{NULL}.}
}
\value{
[\code{smoof_multi_objective_function}]
}
\description{
Generates an NK-landscapes with input dimension \code{N}, output dimension \code{M}
and epistatic links/interactions between bits specified by the vector \code{K}.
The correlation between the objectives can be adjusted via the correlation
parameter \eqn{\rho \in [-1,1]}.
}
\examples{
# generate homogeneous uncorrelated bi-objective rMNK-landscape with each
# three epistatic links
M = 2L
N = 20L
K = 3L
fn = makeRMNKFunction(M, N, K)

# generate strongly correlated objectives
fn = makeRMNKFunction(rho = 0.9, M, N, K)

# generate first function has 3 epistatic links while the second has 2
fn = makeRMNKFunction(rho = 0.9, M, N, K = c(3L, 2L))

# alternative constructor: generate two single-objective NK-landscapes
# and combine into bi-objective problem
soofn1 = makeNKFunction(N, K) # homegeneous in K
K = sample(2:3, size = N, replace = TRUE)
soofn2 = makeNKfunction(N, K = K) # heterogeneous in K
moofn = makeRMNKFunction(funs = list(soofn1, soofn2))
getNumberOfObjectives(moofn)

}
\seealso{
\code{\link{makeNKFunction}}

Other nk_landscapes: 
\code{\link{exportNKFunction}()},
\code{\link{makeNKFunction}()}
}
\concept{nk_landscapes}
