% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.rosenbrock.R
\name{makeRosenbrockFunction}
\alias{makeRosenbrockFunction}
\title{Rosenbrock function}
\usage{
makeRosenbrockFunction(dimensions)
}
\description{
Also known as the second function of De Jong or the (Rosenbrock) banana function.
}

