% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.schaffer.function.2.R
\name{makeSchafferN2Function}
\alias{makeSchafferN2Function}
\title{Modified Schaffer Function N. 2}
\usage{
makeSchafferN2Function()
}
\value{
An object of class \code{SingleObjectiveFunction}, representing the Modified Schaffer N. 2 Function.

[\code{smoof_single_objective_function}]
}
\description{
Second function by Schaffer. The definition is given by the formula
\deqn{f(\mathbf{x}) = 0.5 + \frac{\sin^2(\mathbf{x}_1^2 - \mathbf{x}_2^2) - 0.5}{(1 + 0.001(\mathbf{x}_1^2 + \mathbf{x}_2^2))^2}}
subject to \eqn{\mathbf{x}_i \in [-100, 100], i = 1, 2}.
}
\references{
S. K. Mishra, Some New Test Functions For Global Optimization
And Performance of Repulsive Particle Swarm Method.
}
