% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.schaffer.function.4.R
\name{makeSchafferN4Function}
\alias{makeSchafferN4Function}
\title{Schaffer Function N. 4}
\usage{
makeSchafferN4Function()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Second function by Schaffer. The defintion is given by the formula
\deqn{f(\mathbf{x}) = 0.5 + \frac{\cos^2(sin(|\mathbf{x}_1^2 - \mathbf{x}_2^2|)) - 0.5}{(1 + 0.001(\mathbf{x}_1^2 + \mathbf{x}_2^2))^2}}
subject to \eqn{\mathbf{x}_i \in [-100, 100], i = 1, 2}.
}
\references{
S. K. Mishra, Some New Test Functions For Global Optimization
And Performance of Repulsive Particle Swarm Method.
}

