% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.schwefel.R
\name{makeSchwefelFunction}
\alias{makeSchwefelFunction}
\title{Schwefel function}
\usage{
makeSchwefelFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{otf_single_objective_function}]
}
\description{
Schwefel function
}

