% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sof.shubert.R
\name{makeShubertFunction}
\alias{makeShubertFunction}
\title{Shubert Function}
\usage{
makeShubertFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
The defintion of this two-dimensional function is given by
\deqn{f(\mathbf{x}) = \prod_{i = 1}^{2} \left(\sum_{j = 1}^{5} \cos((j + 1)\mathbf{x}_i + j\right)}
subject to \eqn{\mathbf{x}_i \in [-10, 10], i = 1, 2}.
}
\references{
J. P. Hennart (ed.), Numerical Analysis, Proc. 3rd AS Workshop,
Lecture Notes in Mathematics, vol. 90, Springer, 1982.
}

