% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeSingleObjectiveFunction}
\alias{makeSingleObjectiveFunction}
\title{Generator for function of otf type.}
\usage{
makeSingleObjectiveFunction(name, fn, par.set, noisy = FALSE,
  constraint.fn = NULL, global.opt.params = NULL, global.opt.value = NULL)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
Function name.}

\item{fn}{[\code{function}]\cr
Target function.}

\item{par.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
Parameter set describing different ascpects of the target function parameters, i. e.,
names, lower and/or upper bounds, types and so on. See \code{\link[ParamHelpers]{makeParamSet}}
for further information.}

\item{noisy}{[\code{logical(1)}]\cr
Is the function noisy? Defaults to \code{FALSE}.}

\item{constraint.fn}{[\code{function | NULL}]\cr
Function which returns a logical vector indicating which indicates whether certain conditions
are met or not. Default is \code{NULL}, which means, that there are no constraints (beside possible)
box constraints.}

\item{global.opt.params}{[\code{list}]\cr
List of named parameter values of the global optimum. Default is \code{NULL} which means unknown.}

\item{global.opt.value}{[\code{numeric(1)}]\cr
Global optimum value if known. Default is \code{NULL}.}
}
\value{
[\code{function}] Target function with additional stuff attached as attributes.
}
\description{
Generator for function of otf type.
}

