% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.sphere.R
\name{makeSphereFunction}
\alias{makeSphereFunction}
\title{Sphere function}
\usage{
makeSphereFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Also known as the the \dQuote{De Jong function 1}. Convex, continous function
calculated via the formula \deqn{f(\mathbf{x}) = \sum_{i=1}^{n} \mathbf{x}_i.}
}

