% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.step.n1.R
\name{makeStepN1Function}
\alias{makeStepN1Function}
\title{Step function N. 1}
\usage{
makeStepN1Function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
This function is based on the defintion
\deqn{f(\mathbf{x}) = \sum_{i = 1}^{n} (\lfloor |\mathbf{x}_i|\rfloor)}
subject to \eqn{\mathbf{x}_i \in [-100, 100]} for \eqn{i = 1, \ldots, n}.
}

