% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.step.n2.R
\name{makeStepN2Function}
\alias{makeStepN2Function}
\title{Step function N. 2}
\usage{
makeStepN2Function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
This function is based on the defintion
\deqn{f(\mathbf{x}) = \sum_{i = 1}^{n} (\lfloor\mathbf{x}_i + 0.5\rfloor)^2}
subject to \eqn{\mathbf{x}_i \in [-100, 100]} for \eqn{i = 1, \ldots, n}.
}
\references{
T. Baeck, H. P. Schwefel, An Overview of Evolutionary Algorithm
for Parameter Optimization, Evolutionary Computation, vol. 1, no. 1, pp. 1-23, 1993.
}

