% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.step.n3.R
\name{makeStepN3Function}
\alias{makeStepN3Function}
\title{Step function N. 3}
\usage{
makeStepN3Function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
This function is based on the defintion
\deqn{f(\mathbf{x}) = \sum_{i = 1}^{n} (\lfloor \mathbf{x}_i^2\rfloor)}
subject to \eqn{\mathbf{x}_i \in [-100, 100]} for \eqn{i = 1, \ldots, n}.
}

