% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.styblinski.tang.R
\name{makeStyblinkskTangFunction}
\alias{makeStyblinkskTangFunction}
\title{Styblinksk-Tang function}
\usage{
makeStyblinkskTangFunction(dimensions)
}
\description{
Styblinksk-Tang function
}

