% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.sum.of.different.powers.R
\name{makeSumOfDifferentSquaresFunction}
\alias{makeSumOfDifferentSquaresFunction}
\title{Sum of different squares function}
\usage{
makeSumOfDifferentSquaresFunction(dimensions)
}
\description{
Sum of different squares function
}

