% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.swiler2014.R
\name{makeSwiler2014Function}
\alias{makeSwiler2014Function}
\title{Swiler2014 function.}
\usage{
makeSwiler2014Function()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Mixed parameter space with one discrete parameter \eqn{x_1 \in \{1, 2, 3, 4, 5\}}
and two numerical parameters \eqn{x_1, x_2 \in [0, 1]}. The function is defined
as follows:
\deqn{
f(\mathbf{x}) = \\
\sin(2\pi x_3 - \pi) + 7 \sin^2(2 \pi x_2 - \pi) \, if \, x_1 = 1 \\
\sin(2\pi x_3 - \pi) + 7 \sin^2(2 \pi x_2 - \pi) + 12 \sin(2 \pi x_3 - \pi) \, if \, x_1 = 2 \\
\sin(2\pi x_3 - \pi) + 7 \sin^2(2 \pi x_2 - \pi) + 0.5 \sin(2 \pi x_3 - \pi) \, if \, x_1 = 3 \\
\sin(2\pi x_3 - \pi) + 7 \sin^2(2 \pi x_2 - \pi) + 8.0 \sin(2 \pi x_3 - \pi) \, if \, x_1 = 4 \\
\sin(2\pi x_3 - \pi) + 7 \sin^2(2 \pi x_2 - \pi) + 3.5 \sin(2 \pi x_3 - \pi) \, if \, x_1 = 5.
}
}
\references{
See \url{https://cfwebprod.sandia.gov/cfdocs/CompResearch/docs/MixedSurrSAND2012-0491.pdf}.
}

