% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.three.hump.camel.R
\name{makeThreeHumpCamelFunction}
\alias{makeThreeHumpCamelFunction}
\title{Three-Hump Camel Function}
\usage{
makeThreeHumpCamelFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Three-Hump Camel Function
}

