% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.trecanni.R
\name{makeTrecanniFunction}
\alias{makeTrecanniFunction}
\title{Trecanni Function}
\usage{
makeTrecanniFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Trecanni Function
}

