% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUFFunction.R
\name{makeUFFunction}
\alias{makeUFFunction}
\title{Generator for the functions UF1, ..., UF10 of the CEC 2009.}
\usage{
makeUFFunction(dimensions, id)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}

\item{id}{[\code{integer(1)}]\cr
Instance identifier. Integer value between 1 and 10.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Generator for the functions UF1, ..., UF10 of the CEC 2009.
}
\note{
The implementation is based on the original CPP implemenation by Qingfu
Zhang, Aimin Zhou, Shizheng Zhaoy, Ponnuthurai Nagaratnam Suganthany, Wudong Liu
and Santosh Tiwar.
}
\author{
Jakob Bossek \email{j.bossek@gmail.com}
}
\references{
See the \href{http://web.mysites.ntu.edu.sg/epnsugan/PublicSite/Shared\%20Documents/CEC2009-MOEA/}{CEC2009-MOEA}
for a detailed description of the functions.
}

