% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.zettl.R
\name{makeZettlFunction}
\alias{makeZettlFunction}
\title{Zettl Function}
\usage{
makeZettlFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
The unimodal Zettl Function is based on the defintion
\deqn{f(\mathbf{x}) = (\mathbf{x}_1^2 + \mathbf{x}_2^2 - 2\mathbf{x}_1)^2 + 0.25 \mathbf{x}_1}
with box-constraints \eqn{\mathbf{x}_i \in [-5, 10], i = 1, 2}.
}
\references{
H. P. Schwefel, Evolution and Optimum Seeking, John Wiley Sons, 1995.
}

