% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveOrigin.R
\name{moveOrigin}
\alias{moveOrigin}
\title{Return a function with a shifted origin.}
\usage{
moveOrigin(fn, shift)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Smoof function which should be shifted}

\item{shift}{[\code{numeric}]\cr
Location of origin of the shifted function in the parameter space of \code{fn}}
}
\value{
[\code{smoof_shifted_function}]
}
\description{
Shift the parameter space to place the shift in the origin of the shifted function that is returned.
This is useful to test if an algorithm is sensitive to the absolute location of the global optimum.

Note that all other aspects of the function are also shifted (global and local optima, and any box constraints).
Therefore the relative location of the optima w.r.t. the bounding box do not change, only the absolute position!
This is different from the shift transformation applied by BBOB to generate different instances.
}
\examples{
fn = makeBBOBFunction(dimensions = 2L, fid = 1L, iid = 1L)
sfn = moveOrigin(fn, c(2, 1))

# New origin is at (2, 1) in the original parameter space
sfn(c(0, 0)) == fn(c(2, 1))

}
