% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.plot.R
\name{plot.smoof_function}
\alias{plot.smoof_function}
\title{Generate \code{\link[ggplot2]{ggplot}} object.}
\usage{
\method{plot}{smoof_function}(x, ...)
}
\arguments{
\item{x}{[\code{smoof_function}]\cr
Function.}

\item{...}{[any]\cr
Further parameters passed to the corresponding plot functions.}
}
\value{
Nothing
}
\description{
Generate \code{\link[ggplot2]{ggplot}} object.
}

