% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.plot.R
\name{plot2DNumeric}
\alias{plot2DNumeric}
\title{Plot a two-dimensional numeric function.}
\usage{
plot2DNumeric(x, show.optimum = FALSE, render.levels = FALSE,
  render.contours = TRUE, n.samples = 100L, ...)
}
\arguments{
\item{x}{[\code{smoof_function}]\cr
Function.}

\item{show.optimum}{[\code{logical(1)}]\cr
If the function has a known global optimum, should its location be
plotted by a point or multiple points in case of multiple global optima?
Default is \code{FALSE}.}

\item{render.levels}{[\code{logical(1)}]\cr
Show a level-plot? Default is \code{FALSE}.}

\item{render.contours}{[\code{logical(1)}]\cr
Render contours? Default is \code{TRUE}.}

\item{n.samples}{[\code{integer(1)}]\cr
Number of locations per dimension to be sampled. Default is 100.}

\item{...}{[any]\cr
Further paramerters passed to \code{image} respectively \code{contour} function.}
}
\value{
Nothing
}
\description{
Either a contour-plot or a level-plot.
}

