% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resetEvaluationCounter.R
\name{resetEvaluationCounter}
\alias{resetEvaluationCounter}
\title{Reset evaluation counter.}
\usage{
resetEvaluationCounter(fn)
}
\arguments{
\item{fn}{[\code{smoof_counting_function}]\cr
Wrapped \code{smoof_function}.}
}
\description{
Reset evaluation counter.
}

