% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shouldBeMinimized.R
\name{shouldBeMinimized}
\alias{shouldBeMinimized}
\title{Check if function should be minimized.}
\usage{
shouldBeMinimized(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{logical}] Each component indicates whether the corresponding
objective should be minimized.
}
\description{
Functions can have an associated global optimum. In this case
one needs to know whether the optimum is a minimum or a maximum.
}
