% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addLoggingWrapper.R
\name{addLoggingWrapper}
\alias{addLoggingWrapper}
\title{Return a function which internally stores x or y values.}
\usage{
addLoggingWrapper(fn, logg.x = FALSE, logg.y = TRUE)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Smoof function.}

\item{logg.x}{[\code{logical(1)}]\cr
Should x-values be logged? Default is \code{FALSE}.}

\item{logg.y}{[\code{logical(1)}]\cr
Should objective values be logged? Default is \code{TRUE}.}
}
\value{
[\code{smoof_logging_function}]
}
\description{
Return a function which internally stores x or y values.
}

