% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.nk.R
\name{exportNKFunction}
\alias{exportNKFunction}
\alias{importNKFunction}
\title{Export/import (rM)NK-landscapes}
\usage{
exportNKFunction(x, path)

importNKFunction(path)
}
\arguments{
\item{x}{[\code{smoof_function}]\cr
NK-landscape of rMNK-landscape.}

\item{path}{[\code{character(1)}]\cr
Path to file.}
}
\value{
Silently returns \code{TRUE} on success.
}
\description{
NK-landscapes and rMNK-landscapes are randomly generated combinatorial structures.
In contrast to continuous benchmark function it thus makes perfect sense to store
the landscape definitions in a text-based format.
}
\details{
Note: the function overwrites existing files without asking.
}
\seealso{
importNKFunction

Other nk_landscapes: 
\code{\link{makeNKFunction}()},
\code{\link{makeRMNKFunction}()}
}
\concept{nk_landscapes}
