% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterFunctionByTags.R
\name{filterFunctionsByTags}
\alias{filterFunctionsByTags}
\title{Get a list of implemented test functions with specific tags.}
\usage{
filterFunctionsByTags(tags)
}
\arguments{
\item{tags}{[\code{character}]\cr
Character vector of tags.}
}
\value{
[\code{character}]
  Named vector of function names with the given tags.
}
\description{
Get a list of implemented test functions with specific tags.
}

