% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAvailableTags.R
\name{getAvailableTags}
\alias{getAvailableTags}
\title{Returns a character vector of possible function tags.}
\usage{
getAvailableTags()
}
\value{
[\code{character}]
}
\description{
Test function are frequently distinguished by characteristic high-level properties,
e.g., unimodal or multimodal, continuous or discontinuous, separable or non-separable.
The \pkg{smoof} package offers the possibility to associate a set of properties,
termed \dQuote{tags} to a \code{smoof_function}. This helper function returns
a character vector of all possible tags.
}

