% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAvailableTags.R
\name{getAvailableTags}
\alias{getAvailableTags}
\title{Returns a character vector of possible function tags.}
\usage{
getAvailableTags()
}
\value{
[\code{character}]
}
\description{
Returns a character vector of possible function tags.
}

