% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDescription.R
\name{getDescription}
\alias{getDescription}
\title{Return description of the function.}
\usage{
getDescription(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{character(1)}]
}
\description{
Return description of the function.
}

