% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGlobalOptimum}
\alias{getGlobalOptimum}
\title{Returns the global optimum and its value.}
\usage{
getGlobalOptimum(fn)
}
\arguments{
\item{fn}{[\code{otf_function}]\cr
Target function.}
}
\value{
[\code{list}] List containing the following entries:
\itemize{
  \item{param [\code{list}]}{Named list of parameter value(s).}
  \item{value [\code{numeric(1)}]}{Optimal value.}
}
}
\description{
Returns the global optimum and its value.
}
\note{
Keep in mind, that this method makes sense only for single-objective target function.
}

