% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getID.R
\name{getID}
\alias{getID}
\title{Return the ID / short name of the function.}
\usage{
getID(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{character(1)}]
}
\description{
Return the ID / short name of the function.
}

