% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoxConstraints.R
\name{getLowerBoxConstraints}
\alias{getLowerBoxConstraints}
\title{Return lower box constaints.}
\usage{
getLowerBoxConstraints(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{numeric}]
}
\description{
Return lower box constaints.
}

