% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getName.R
\name{getName}
\alias{getName}
\title{Return the name of the function.}
\usage{
getName(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{character(1)}]
}
\description{
Return the name of the function.
}

