% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumberOfEvaluations.R
\name{getNumberOfEvaluations}
\alias{getNumberOfEvaluations}
\title{Returns the number of function evaluations performed by the wrapped
\code{smoof_function}.}
\usage{
getNumberOfEvaluations(fn)
}
\arguments{
\item{fn}{[\code{smoof_counting_function}]\cr
Wrapped \code{smoof_function}.}
}
\value{
[\code{integer(1)}]
 The number of function evaluations.
}
\description{
Returns the number of function evaluations performed by the wrapped
\code{smoof_function}.
}
