% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getNumberOfObjectives}
\alias{getNumberOfObjectives}
\title{Determine the number of objectives.}
\usage{
getNumberOfObjectives(fn)
}
\arguments{
\item{fn}{[\code{otf_function}]\cr
Objective function.}
}
\value{
[\code{integer(1)}]
}
\description{
Determine the number of objectives.
}

