% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getNumberOfParameters}
\alias{getNumberOfParameters}
\title{Determine the number of parameters.}
\usage{
getNumberOfParameters(fn)
}
\arguments{
\item{fn}{[\code{otf_function}]\cr
Target function.}
}
\value{
[\code{integer(1)}]
}
\description{
Determine the number of parameters.
}

