% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNumberOfParameters.R
\name{getNumberOfParameters}
\alias{getNumberOfParameters}
\title{Determine the number of parameters.}
\usage{
getNumberOfParameters(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{integer(1)}]
}
\description{
Determine the number of parameters.
}

