% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getParamSet.R
\name{getParamSet}
\alias{getParamSet}
\title{Get parameter set.}
\usage{
getParamSet(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{\link[ParamHelpers]{ParamSet}}]
}
\description{
Each smoof function contains a parameter set of type \code{\link[ParamHelpers]{ParamSet}}
assigned to it, which describes types and bounds of the function parameters.
This function returns the parameter set.
}
\examples{
fn = makeSphereFunction(3L)
ps = smoof::getParamSet(fn)
print(ps)
}

