% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getTags.R
\name{getTags}
\alias{getTags}
\title{Returns vector of associated tags.}
\usage{
getTags(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Target function.}
}
\value{
[\code{character}]
}
\description{
Returns vector of associated tags.
}

