% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getWrappedFunction.R
\name{getWrappedFunction}
\alias{getWrappedFunction}
\title{Get the wrapped function. Useful to get the smoof function, if it is wrapped
with e.g. a counting wrapper.}
\usage{
getWrappedFunction(fn)
}
\arguments{
\item{fn}{[\code{smoof_wrapped_function}]\cr
Wrapping function.}
}
\value{
[\code{function}]
}
\description{
Get the wrapped function. Useful to get the smoof function, if it is wrapped
with e.g. a counting wrapper.
}

