% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hasConstraints.R
\name{hasConstraints}
\alias{hasConstraints}
\title{Checks whether the objective function has constraints.}
\usage{
hasConstraints(fn)
}
\arguments{
\item{fn}{[\code{otf_function}]\cr
Target function.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether the objective function has constraints.
}

