% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasGlobalOptimum.R
\name{hasGlobalOptimum}
\alias{hasGlobalOptimum}
\title{Checks whether global optimum is known.}
\usage{
hasGlobalOptimum(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether global optimum is known.
}
