% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getGlobalOptimum.R
\name{hasGlobalOptimum}
\alias{hasGlobalOptimum}
\title{Checks whether global optimum is known.}
\usage{
hasGlobalOptimum(fn)
}
\arguments{
\item{fn}{[\code{otf_function}]\cr
Target function.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether global optimum is known.
}

