% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hasTags.R
\name{hasTags}
\alias{hasTags}
\title{Check if function has assigend special tags.}
\usage{
hasTags(fn, tags)
}
\arguments{
\item{fn}{[\code{smoof_function}]
Function of \code{smoof_function}, a \code{smoof_generator} or a string.}

\item{tags}{[\code{character}]\cr
Vector of tags/properties to check \code{fn} for.}
}
\value{
[\code{logical(1)}]
}
\description{
Check if function has assigend special tags.
}

