% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isMultiobjective.R
\name{isMultiobjective}
\alias{isMultiobjective}
\title{Checks whether the given function is multi-objective.}
\usage{
isMultiobjective(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{logical(1)}] \code{TRUE} if function is multi-objective.
}
\description{
Checks whether the given function is multi-objective.
}

