% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNoisy.R
\name{isNoisy}
\alias{isNoisy}
\title{Checks whether the given function is noisy.}
\usage{
isNoisy(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{logical(1)}]
 \code{TRUE} if the function is noisy, \code{FALSE} otherwise.
}
\description{
Checks whether the given function is noisy.
}
