% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isNoisy.R
\name{isNoisy}
\alias{isNoisy}
\title{Checks whether the given function is noisy.}
\usage{
isNoisy(fn)
}
\arguments{
\item{fn}{[\code{otf_function}]\cr
Target function.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether the given function is noisy.
}

