% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isOTFFunction}
\alias{isOTFFunction}
\title{Checks whether the given object is a otf_function.}
\usage{
isOTFFunction(object)
}
\arguments{
\item{object}{[any]\cr
Arbitrary R object.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether the given object is a otf_function.
}

