% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSingleobjective.R
\name{isSingleobjective}
\alias{isSingleobjective}
\title{Checks whether the given function is single-objective.}
\usage{
isSingleobjective(fn)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}
}
\value{
[\code{logical(1)}] \code{TRUE} if function is single-objective.
}
\description{
Checks whether the given function is single-objective.
}

