% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isOTFFunction.R
\name{issmoofFunction}
\alias{issmoofFunction}
\title{Checks whether the given object is a smoof_function.}
\usage{
issmoofFunction(object)
}
\arguments{
\item{object}{[any]\cr
Arbitrary R object.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether the given object is a smoof_function.
}

