% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.ackley.R
\name{makeAckleyFunction}
\alias{makeAckleyFunction}
\title{Ackley function.}
\usage{
makeAckleyFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Ackley function.
}

