% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.alpine01.R
\name{makeAlpine01Function}
\alias{makeAlpine01Function}
\title{Alpine01 function}
\usage{
makeAlpine01Function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
An object of class \code{SingleObjectiveFunction}, representing the Alpine01 Function.

[\code{smoof_single_objective_function}]
}
\description{
Highly multi-modal single-objective optimization test function. It is defined
as \deqn{f(\mathbf{x}) = \sum_{i = 1}^{n} |\mathbf{x}_i \sin(\mathbf{x}_i) + 0.1\mathbf{x}_i|}
with box constraints \eqn{\mathbf{x}_i \in [-10, 10]} for \eqn{i = 1, \ldots, n}.
}
\references{
S. Rahnamyan, H. R. Tizhoosh, N. M. M. Salama, A Novel Population
Initialization Method for Accelerating Evolutionary Algorithms, Computers and
Mathematics with Applications, vol. 53, no. 10, pp. 1605-1614, 2007.
}
