% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.alpine01.R
\name{makeAlpine01Function}
\alias{makeAlpine01Function}
\title{Alpine01 function}
\usage{
makeAlpine01Function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{otf_single_objective_function}]
}
\description{
Alpine01 function
}

