% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.amgm.R
\name{makeAMGMFunction}
\alias{makeAMGMFunction}
\title{AMGM function}
\usage{
makeAMGMFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Aritmetic Mean - Geometric Mean Equality problem.
}

