% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.bukin.n2.R
\name{makeBukinN2Function}
\alias{makeBukinN2Function}
\title{Bukin function N. 2}
\usage{
makeBukinN2Function()
}
\value{
[\code{otf_single_objective_function}]
}
\description{
Bukin function N. 2
}

