% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.bukin.n4.R
\name{makeBukinN4Function}
\alias{makeBukinN4Function}
\title{Bukin function N. 4}
\usage{
makeBukinN4Function()
}
\value{
[\code{otf_single_objective_function}]
}
\description{
Bukin function N. 4
}

