% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.cosine.mixture.R
\name{makeCosineMixtureFunction}
\alias{makeCosineMixtureFunction}
\title{Cosine Mixture Function}
\usage{
makeCosineMixtureFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
An object of class \code{SingleObjectiveFunction}, representing the Cosine Mixture Function.

[\code{smoof_single_objective_function}]
}
\description{
Single-objective test function based on the formula
\deqn{f(\mathbf{x}) = -0.1 \sum_{i = 1}^{n} \cos(5\pi\mathbf{x}_i) - \sum_{i = 1}^{n} \mathbf{x}_i^2}
subject to \eqn{\mathbf{x}_i \in [-1, 1]} for \eqn{i = 1, \ldots, n}.
}
\references{
M. M. Ali, C. Khompatraporn, Z. B. Zabinsky, A Numerical Evaluation
of Several Stochastic Algorithms on Selected Continuous Global Optimization
Test Problems, Journal of Global Optimization, vol. 31, pp. 635-672, 2005.
}
