% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.crowned.cross.R
\name{makeCrownedCrossFunction}
\alias{makeCrownedCrossFunction}
\title{CrownedCross Function}
\usage{
makeCrownedCrossFunction()
}
\value{
[\code{otf_single_objective_function}]
}
\description{
See also CrossInTray function.
}

