% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.damavandi.R
\name{makeDamavandiFunction}
\alias{makeDamavandiFunction}
\title{Damavandi function}
\usage{
makeDamavandiFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Damavandi function
}

