% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.decanomial.R
\name{makeDecanomialFunction}
\alias{makeDecanomialFunction}
\title{Decanomial function}
\usage{
makeDecanomialFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Decanomial function
}

