% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.deflected.corrugated.spring.R
\name{makeDeflectedCorrugatedSpringFunction}
\alias{makeDeflectedCorrugatedSpringFunction}
\title{Deflected Corrugated Spring function}
\usage{
makeDeflectedCorrugatedSpringFunction(dimensions, K = 5, alpha = 5)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}

\item{K}{[\code{numeric(1)}]\cr
Parameter.
Default is 5.}

\item{alpha}{[\code{numeric(1)}]\cr
Parameter.
Default is 5.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Scalable single-objective test function based on the formula
\deqn{f(\mathbf{x}) = 0.1 \sum_{i = 1}^{n} (x_i - \alpha)^2 - \cos\left(K \sqrt{\sum_{i = 1}^{n} (x_i - \alpha)^2}\right)}
with \eqn{\mathbf{x}_i \in [0, 2\alpha], i = 1, \ldots, n} and \eqn{\alpha = K = 5}
by default.
}
\references{
See \url{https://al-roomi.org/benchmarks/unconstrained/n-dimensions/238-deflected-corrugated-spring-function}.
}
