% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.deflected.corrugated.spring.R
\name{makeDeflectedCorrugatedSpringFunction}
\alias{makeDeflectedCorrugatedSpringFunction}
\title{Deflected Corrugated Spring function}
\usage{
makeDeflectedCorrugatedSpringFunction(dimensions, K = 5, alpha = 5)
}
\description{
Deflected Corrugated Spring function
}

