% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.double.sum.R
\name{makeDoubleSumFunction}
\alias{makeDoubleSumFunction}
\title{Double-Sum function.}
\usage{
makeDoubleSumFunction(dimensions)
}
\description{
Also known as the rotated hyper-ellipsoid function.
}

