% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mof.dtlz1.R
\name{makeDTLZ1Function}
\alias{makeDTLZ1Function}
\title{DTLZ1 Function (family)}
\usage{
makeDTLZ1Function(dimensions, n.objectives)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Number of decision variables.}

\item{n.objectives}{[\code{integer(1)}]\cr
Number of objectives.}
}
\value{
[\code{smoof_multi_objective_function}]
}
\description{
Builds and returns the multi-objective DTLZ1 test problem.
}
\references{
K. Deb and L. Thiele and M. Laumanns and E. Zitzler. Scalable
Multi-Objective Optimization Test Problems.
}

