% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mof.dtlz6.R
\name{makeDTLZ6Function}
\alias{makeDTLZ6Function}
\title{DTLZ6 Function (family)}
\usage{
makeDTLZ6Function(dimensions, n.objectives)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Number of decision variables.}

\item{n.objectives}{[\code{integer(1)}]\cr
Number of objectives.}
}
\value{
[\code{smoof_multi_objective_function}]
}
\description{
Builds and returns the multi-objective DTLZ6 test problem. This problem
can be characterized by a disconnected Pareto-optimal front in the search
space. This introduces a new challenge to evolutionary multi-objective
optimizers, i.e., to maintain different subpopulations within the search
space to cover the entire Pareto-optimal front.
\deqn{\text{Minimize } &\; f_1(\textbf{x}) = (1+g(\textbf{x}_M)) \cos(\theta_1\pi/2) \cos(\theta_2\pi/2) \cdots \cos(\theta_{M-2}\pi/2) \cos(\theta_{M-1}\pi/2),\\
\text{Minimize } &\; f_2(\textbf{x}) = (1+g(\textbf{x}_M)) \cos(\theta_1\pi/2) \cos(\theta_2\pi/2) \cdots \cos(\theta_{M-2}\pi/2) \sin(\theta_{M-1}\pi/2),\\
\text{Minimize } &\; f_3(\textbf{x}) = (1+g(\textbf{x}_M)) \cos(\theta_1\pi/2) \cos(\theta_2\pi/2) \cdots \sin(\theta_{M-2}\pi/2),\\
\vdots &\; \vdots\\
\text{Minimize } &\; f_{M-1}(\textbf{x}) = (1+g(\textbf{x}_M)) \cos(\theta_1\pi/2) \sin(\theta_2\pi/2),\\
\text{Minimize } &\; f_{M}(\textbf{x}) = (1+g(\textbf{x}_M)) \sin(\theta_1\pi/2),\\
&\; 0 \leq x_i \leq 1, \text{ for } i=1,2,\dots,n,\\
\text{where} &\; \theta_i = \frac{\pi}{4(1+ g)} (1+2gx_i), \text{ for } i = 2,3,\dots,(M-1)\\
&\; g(\textbf{x}_M) = \sum\limits_{x_i\in\textbf{x}_M}x_i^{0.1}.}
}
\references{
K. Deb and L. Thiele and M. Laumanns and E. Zitzler. Scalable
Multi-Objective Optimization Test Problems. Computer Engineering and Networks
Laboratory (TIK), Swiss Federal Institute of Technology (ETH) Zurich, 112, 2001
}

