% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sof.eggcrate.R
\name{makeEggCrateFunction}
\alias{makeEggCrateFunction}
\title{Egg Crate Function}
\usage{
makeEggCrateFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
This single-objective function follows the definition
\deqn{f(\mathbf{x}) = \mathbf{x}_1^2 + \mathbf{x}_2^2 + 25(\sin^2(\mathbf{x}_1) + \sin^2(\mathbf{x}_2))}
with \eqn{\mathbf{x}_i \in [-5, 5]} for \eqn{i = 1, 2}.
}

