% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sof.el.attar.vidyasagar.dutta.R
\name{makeElAttarVidyasagarDuttaFunction}
\alias{makeElAttarVidyasagarDuttaFunction}
\title{El-Attar-Vidyasagar-Dutta Function}
\usage{
makeElAttarVidyasagarDuttaFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
This function is based on the formula
\deqn{f(\mathbf{x}) = (\mathbf{x}_1^2 + \mathbf{x}_2 - 10)^2 + (\mathbf{x}_1 + \mathbf{x}_2^2 - 7)^2 + (\mathbf{x}_1^2 + \mathbf{x}_2^3 - 1)^2}
subject to \eqn{\mathbf{x}_i \in [-500, 500], i = 1, 2}.
}
\references{
R. A. El-Attar, M. Vidyasagar, S. R. K. Dutta, An Algorithm for
II-norm Minimiza- tion With Application to Nonlinear II-approximation, SIAM
Journal on Numverical Analysis, vol. 16, no. 1, pp. 70-86, 1979.
}

