% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sof.freudenstein.roth.R
\name{makeFreudensteinRothFunction}
\alias{makeFreudensteinRothFunction}
\title{Freudenstein Roth Function}
\usage{
makeFreudensteinRothFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
This test function is based on the formula
\deqn{f(\mathbf{x}) = (\mathbf{x}_1 - 13 + ((5 - \mathbf{x}_2)\mathbf{x}_2 - 2)\mathbf{x}_2)^2 + (\mathbf{x}_1 - 29 + ((\mathbf{x}_2 + 1)\mathbf{x}_2 - 14)\mathbf{x}_2)^2}
subject to \eqn{\mathbf{x}_i \in [-10, 10], i = 1, 2}.
}
\references{
S. S. Rao, Engineering Optimization: Theory and Practice,
John Wiley & Sons, 2009.
}

