% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.goldstein.price.R
\name{makeGoldsteinPriceFunction}
\alias{makeGoldsteinPriceFunction}
\title{Goldstein-Price Function}
\usage{
makeGoldsteinPriceFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Two-dimensional test function for global optimization. The implementation
follows the formula:
\deqn{f(\mathbf{x}) = \left(1 + (\mathbf{x}_1 + \mathbf{x}_2 + 1)^2 \cdot (19 - 14\mathbf{x}_1 + 3\mathbf{x}_1^2 - 14\mathbf{x}_2 + 6\mathbf{x}_1\mathbf{x}_2 + 3\mathbf{x}_2^2)\right)\\ \qquad \cdot \left(30 + (2\mathbf{x}_1 - 3\mathbf{x}_2)^2 \cdot (18 - 32\mathbf{x}_1 + 12\mathbf{x}_1^2 + 48\mathbf{x}_2 - 36\mathbf{x}_1\mathbf{x}_2 + 27\mathbf{x}_2^2)\right)}
with \eqn{\mathbf{x}_i \in [-2, 2], i = 1, 2}.
}
\references{
Goldstein, A. A. and Price, I. F.: On descent from local minima.
Math. Comput., Vol. 25, No. 115, 1971.
}

