% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.himmelblau.R
\name{makeHimmelblauFunction}
\alias{makeHimmelblauFunction}
\title{Himmelblau Function}
\usage{
makeHimmelblauFunction()
}
\value{
[\code{otf_single_objective_function}]
}
\description{
Himmelblau Function
}

