% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.hyper.ellipsoid.R
\name{makeHyperEllipsoidFunction}
\alias{makeHyperEllipsoidFunction}
\title{Hyper-Ellipsoid function}
\usage{
makeHyperEllipsoidFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Unimodal, convex test function similar to the Sphere function (see \code{\link{makeSphereFunction}}).
Calculated via the formula: \deqn{f(\mathbf{x}) = \sum_{i=1}^{n} i \cdot \mathbf{x}_i.}
}
