% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.hyper.ellipsoid.R
\name{makeHyperEllipsoidFunction}
\alias{makeHyperEllipsoidFunction}
\title{Hyper-Ellipsoid function}
\usage{
makeHyperEllipsoidFunction(dimensions)
}
\description{
Hyper-Ellipsoid function
}

