% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.mccormick.R
\name{makeMcCormickFunction}
\alias{makeMcCormickFunction}
\title{McCormick Function}
\usage{
makeMcCormickFunction()
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
McCormick Function
}

