% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.michalewicz.R
\name{makeMichalewiczFunction}
\alias{makeMichalewiczFunction}
\title{Michalewicz function}
\usage{
makeMichalewiczFunction(dimensions, m = 5)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}

\item{m}{[\code{integer(1)}]\cr
\dQuote{Steepness} parameter.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Highly multimodal single-objective test function with \eqn{n!} local minima
with the formula:
\deqn{f(\mathbf{x}) = -\sum_{i=1}^{n} \sin(\mathbf{x}_i) \cdot \left(\sin\left(\frac{i \cdot \mathbf{x}_i}{\pi}\right)\right)^{2m}.}
}

