% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.mmf4.R
\name{makeMMF4Function}
\alias{makeMMF4Function}
\title{MMF4 Function}
\usage{
makeMMF4Function()
}
\value{
[\code{smoof_multi_objective_function}]
 Returns an instance of the MMF4 function as a \code{smoof_multi_objective_function} object.
}
\description{
Test problem from the set of "multi-modal multi-objective functions" as for
instance used in the CEC2019 competition.
}
\references{
Caitong Yue, Boyang Qu, Kunjie Yu, Jing Liang, and Xiaodong Li, "A novel
scalable test problem suite for multi-modal multi-objective optimization," in
Swarm and Evolutionary Computation, Volume 48, August 2019, pp. 62–71, Elsevier.
}
