% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.rosenbrock.R
\name{makeRosenbrockFunction}
\alias{makeRosenbrockFunction}
\title{Rosenbrock function}
\usage{
makeRosenbrockFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Also known as the second function of De Jong or the (Rosenbrock) banana function.
}

