% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.schwefel.R
\name{makeSchwefelFunction}
\alias{makeSchwefelFunction}
\title{Schwefel function}
\usage{
makeSchwefelFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Highly multimodal test function. The cursial thing about this function is, that
the global optimum is far away from the next best local optimum.
The function is computed via:
\deqn{f(\mathbf{x}) = \sum_{i=1}^{n} -\mathbf{x}_i \sin\left(\sqrt(|\mathbf{x}_i|)\right)}
with \eqn{\mathbf{x}_i \in [-500, 500], i = 1, \ldots, n.}
}
\references{
Schwefel, H.-P.: Numerical optimization of computer models.
Chichester: Wiley & Sons, 1981.
}

