% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeSingleObjectiveFunction.R
\name{makeSingleObjectiveFunction}
\alias{makeSingleObjectiveFunction}
\title{Generator for single-objective target functions.}
\usage{
makeSingleObjectiveFunction(name, description = NULL, fn,
  has.simple.signature = TRUE, par.set, noisy = FALSE,
  constraint.fn = NULL, tags = character(0), global.opt.params = NULL,
  global.opt.value = NULL)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
Function name. Used for the title of plots for example.}

\item{description}{[\code{character(1)} | \code{NULL}]\cr
Optional function description.}

\item{fn}{[\code{function}]\cr
Target function.}

\item{has.simple.signature}{[\code{logical(1)}]\cr
Set this to \code{TRUE} if the target function expects a vector as input and \code{FALSE}
if it expects a named list of values. The latter is needed if the function depends on mixed
parameters. Default is \code{TRUE}.}

\item{par.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
Parameter set describing different ascpects of the target function parameters, i. e.,
names, lower and/or upper bounds, types and so on. See \code{\link[ParamHelpers]{makeParamSet}}
for further information.}

\item{noisy}{[\code{logical(1)}]\cr
Is the function noisy? Defaults to \code{FALSE}.}

\item{constraint.fn}{[\code{function | NULL}]\cr
Function which returns a logical vector indicating which indicates whether certain conditions
are met or not. Default is \code{NULL}, which means, that there are no constraints (beside possible)
box constraints defined via the \code{par.set} argument.}

\item{tags}{[\code{character}]\cr
Optional character vector of tags or keywords which characterize the function.
E.g. \dQuote{unimodal}, \dQuote{separable}. See \code{\link{getAvailableTags}} for
a list of useful tags.}

\item{global.opt.params}{[\code{list}]\cr
List of named parameter values of the global optimum. Default is \code{NULL} which means unknown.}

\item{global.opt.value}{[\code{numeric(1)}]\cr
Global optimum value if known. Default is \code{NULL}.}
}
\value{
[\code{function}] Target function with additional stuff attached as attributes.
}
\description{
Generator for single-objective target functions.
}
\examples{
library(ggplot2)

  fn = makeSingleObjectiveFunction(
    name = "Sphere Function",
    fn = function(x) sum(x^2),
    par.set = makeNumericParamSet("x", len = 1L, lower = -5L, upper = 5L),
    global.opt.params = list(x = 0)
  )
  print(fn)
  print(autoplot(fn))

  fn.num2 = makeSingleObjectiveFunction(
  	name = "Numeric 2D",
 	fn = function(x) sum(x^2),
 	par.set = makeParamSet(
 		makeNumericParam("x1", lower = -5, upper = 5),
 		makeNumericParam("x2", lower = -5, upper = 5)
 	)
  )
  print(fn)
  print(autoplot(fn))

  fn.mixed = makeSingleObjectiveFunction(
 	name = "Mixed 2D",
 	fn = function(x) x$num1^2 + as.integer(as.character(x$disc1) == "a"),
 	has.simple.signature = FALSE,
 	par.set = makeParamSet(
 		makeNumericParam("num1", lower = -5, upper = 5),
 		makeDiscreteParam("disc1", values = c("a", "b"))
 	),
 	global.opt.params = list(num1 = 0, disc1 = "b")
  )
  print(fn)
  print(autoplot(fn))
}

