% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeSingleObjectiveFunctionFromSOOFunction}
\alias{makeSingleObjectiveFunctionFromSOOFunction}
\title{Generates non-noisy test function from soo function.}
\usage{
makeSingleObjectiveFunctionFromSOOFunction(id, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
Function id of SOO function from package soobench.}

\item{...}{[\code{list}]\cr
Further parameters passed to soo function generator. See the different
soo functions for examples. Probably the most important one is 'dimensions'.}
}
\value{
[\code{otf_function}]
}
\description{
Generates non-noisy test function from soo function.
}
\examples{
\dontrun{
  branin.fn = makeSingleObjectiveFunctionFromSOOFunction("branin")
  print(branin.fn)
}
}

