% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.sphere.R
\name{makeSphereFunction}
\alias{makeSphereFunction}
\title{Sphere Function}
\usage{
makeSphereFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Also known as the the \dQuote{De Jong function 1}. Convex, continous function
calculated via the formula
\deqn{f(\mathbf{x}) = \sum_{i=1}^{n} \mathbf{x}_i^2}
with box-constraings \eqn{\mathbf{x}_i \in [-5.12, 5.12], i = 1, \ldots, n}.
}
\references{
M. A. Schumer, K. Steiglitz, Adaptive Step Size Random Search,
IEEE Transactions on Automatic Control. vol. 13, no. 3, pp. 270-276, 1968.
}
