% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.sum.of.different.powers.R
\name{makeSumOfDifferentSquaresFunction}
\alias{makeSumOfDifferentSquaresFunction}
\title{Sum of Different Squares Function}
\usage{
makeSumOfDifferentSquaresFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Simple unimodal test function similar to the Sphere and Hyper-Ellipsoidal functions.
Formula:
\deqn{f(\mathbf{x}) = \sum_{i=1}^{n} |\mathbf{x}_i|^{i+1}.}
}

