% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.sum.of.different.powers.R
\name{makeSumOfDifferentSquaresFunction}
\alias{makeSumOfDifferentSquaresFunction}
\title{Sum of different squares function}
\usage{
makeSumOfDifferentSquaresFunction(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}]\cr
Size of corresponding parameter space.}
}
\value{
[\code{smoof_single_objective_function}]
}
\description{
Sum of different squares function
}

