% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sof.six.hump.camel.back.function.R, R/sof.three.hump.camel.R
\name{makeThreeHumpCamelFunction}
\alias{makeThreeHumpCamelFunction}
\title{Three-Hump Camel Function}
\usage{
makeThreeHumpCamelFunction()

makeThreeHumpCamelFunction()
}
\description{
Three-Hump Camel Function

Three-Hump Camel Function
}

