% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violatesConstraints.R
\name{violatesConstraints}
\alias{violatesConstraints}
\title{Checks whether constraints are violated.}
\usage{
violatesConstraints(fn, values)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
Objective function.}

\item{values}{[\code{numeric}]\cr
List of values.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks whether constraints are violated.
}
